#39 refactor: User 엔티티를 공통 모듈로 이동하여 구조 정리
- User 엔티티 내 로직은 UserMapper 유틸 클래스로 분리
