+++ yeogi-customer/src/main/java/com/onerty/yeogi/customer/auth/dto/LoginResponse.java
@@ -1,6 +1,6 @@
 package com.onerty.yeogi.customer.auth.dto;
 
-import com.onerty.yeogi.customer.user.User;
+import com.onerty.yeogi.common.user.User;
 
 public record LoginResponse(
         String accessToken,
