+++ yeogi-customer/src/main/java/com/onerty/yeogi/customer/term/AgreementRepository.java
@@ -1,6 +1,6 @@
 package com.onerty.yeogi.customer.term;
 
-import com.onerty.yeogi.customer.user.User;
+import com.onerty.yeogi.common.user.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
