+++ yeogi-customer/src/main/java/com/onerty/yeogi/customer/user/UserService.java
@@ -4,6 +4,7 @@ import com.onerty.yeogi.common.exception.ErrorType;
 import com.onerty.yeogi.common.exception.YeogiException;
 import com.onerty.yeogi.common.term.Term;
 import com.onerty.yeogi.common.term.TermTitle;
+import com.onerty.yeogi.common.user.User;
 import com.onerty.yeogi.customer.term.Agreement;
 import com.onerty.yeogi.customer.term.AgreementId;
 import com.onerty.yeogi.customer.term.AgreementRepository;
@@ -38,7 +39,7 @@ public class UserService {
         validateDuplicateUserAttributes(signupDto.uid(), signupDto.nick());
         validateLatestAndRequiredTerms(signupDto.agreements());
 
-        User user = new User(signupDto);
+        User user = UserMapper.from(signupDto);
         userRepository.save(user);
         saveAgreements(user, signupDto.agreements());
